Solves issues and refactors config
