+++ digital_twins/opc-ua-electroingenium/config.json
@@ -1,0 +1,11 @@
+{
+    "url"           : "opc.tcp://localhost:53530/OPCUA/SimulationServer",
+    "config_ods"    : "configuration.ods",
+    "fmu_filename"  : "Test_DTCONEDAR_windows.fmu",
+    "stop_time"     : 10.0,
+    "step_size"     : 0.5,
+    "record"        : false,
+    "record_interval"   : 5.0,
+    "record_variables"  : [], 
+    "enable_send"       : true
+}
